Base store will now try to use configured AG classes
This is to solve an issue where added a filter to a 'User' class that has the AG::User mixin was not actually used when interacting with the store on the AG side.

The effect was that when a User was registered, and an after_create filter was added to the User model, it never ran because ApiGuardian's UserStore used the ApiGuardian::User model instead.